# Skip host synchronization when it is safe to do so