--- a/src/array.jl
+++ b/src/array.jl
@@ -18,11 +18,11 @@ end
 # indexing
 
 for (f, froc) in (
-        (:blockidx, :blockIdx),
-        (:blockdim, :blockDim),
-        (:threadidx, :threadIdx),
-        (:griddim, :gridGroupDim)
-    )
+    (:blockidx, :blockIdx),
+    (:blockdim, :blockDim),
+    (:threadidx, :threadIdx),
+    (:griddim, :gridGroupDim)
+)
     @eval @inline GPUArrays.$f(::ROCKernelContext) = AMDGPU.$froc().x
 end
 
@@ -47,7 +47,6 @@ end
     return
 end
 
-
 #
 # Host abstractions
 #
@@ -70,7 +69,11 @@ mutable struct ROCArray{T,N} <: AbstractGPUArray{T,N}
     end
 end
 
-_safe_free!(xs::ROCArray) = _safe_free!(xs.buf)
+function _safe_free!(xs::ROCArray)
+    # hip_wait!(xs)
+    _safe_free!(xs.buf)
+end
+
 function _safe_free!(buf::Mem.Buffer)
     Mem.release(buf)
     return