--- a/src/blas/rocBLAS.jl
+++ b/src/blas/rocBLAS.jl
@@ -3,7 +3,7 @@ module rocBLAS
 using ..AMDGPU
 import AMDGPU: hsa_wait!, mark!, librocblas, AnyROCArray
 import AMDGPU: HandleCache, HIP, library_state
-import .HIP: HIPContext, HIPStream, hipContext_t, hipStream_t, hipEvent_t
+import .HIP: HIPContext, HIPStream, HIPEvent, hipContext_t, hipStream_t, hipEvent_t
 
 using LinearAlgebra
 using CEnum