… Project.toml
… src/dnn/batchnorm.jl
… src/dnn/convolution.jl
… src/dnn/pooling.jl
… src/dnn/softmax.jl
… src/highlevel.jl
… src/hip/HIP.jl
… src/hip/libhip.jl
… src/hip/libhip_common.jl
… src/mapreduce.jl
… src/rand/random.jl
… src/rand/rocRAND.jl
… src/random.jl
… src/runtime/kernel.jl
… src/runtime/queue.jl
… src/runtime/sync.jl
… test/rocarray/base.jl
… test/rocarray/nmf.jl
… test/runtests.jl
… test/tls.jl