@pxl-th 2023-05-02
#### Nerf.jl [benchmark](https://github.com/JuliaNeuralGraphics/Nerf.jl/blob/123f638e5d0529bad8b70ea86e19e017025a43f5/src/Nerf.jl#L133) (1000 training steps)
- Before: `103.576595 seconds (18.76 M allocations: 741.180 MiB, 3.55% gc time)`
- After: ` 85.742314 seconds (21.02 M allocations: 782.243 MiB, 1.71% gc time, 0.01% compilation time)`

#### Flux model inference (private repo)
- Before: 60 seconds for an inference
- After: 44 seconds for an inference