@pxl-th 2023-06-16
Superseded by https://github.com/JuliaGPU/AMDGPU.jl/pull/423